Move UITextInput protocol implementation into its own module
