+++ platform/src/os/apple/ios/mod.rs
@@ -2,4 +2,5 @@ pub mod ios;
 pub mod ios_app;
 pub mod ios_event;
 pub mod ios_delegates;
+pub mod ios_text_input;
 pub use self::ios::*;
